+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - the optional `nvim-oxi` and `cmd` attributes to the `nvim_oxi::test` macro
   ([#159](https://github.com/noib3/nvim-oxi/pull/159));
 
+- the optional `library_path` attribute to the `nvim_oxi::test` macro
+  ([#164](https://github.com/noib3/nvim-oxi/pull/164));
+
 ### Changed
 
 - renamed the macro that marks the entrypoint of a plugin from